Dragon can travel (poorly) from user to user
Implemented a light, non-persistent backed that registers sockets as the come in,
and connects them with other sockets as available. When the dragon leaves a user's
screens, the server checks if another user is connected to that edge of the user's
screen. If so, it spawns the dragon anew on the connections' screen. Otherwise it
is respawned on the opposite side of the same user's screen.

The insert and remove logic is naive, connections are basically just glommed on, and could
result in some user's being disconnected if all their connections disappear.