Smarter backend connections graph
Connections and disconnections are handled more intelligently. Connections are
added to an individual in a breadth first manner, and up to four connections are
 added per node.

We remove nodes from the queue when they are deleted.

We don't care so much about directions, now the directions are added randomly
and the direction the dragon exits on doesn't determine which connection the dragon will go to. We can do more deterministic traveling later.