Bug-free adding and removing. Simple graph implementation.
Using an adjacency list graph, and a healthy dose of lodash, created
infrastructure that can add and remove people while maintaining
consistent connectivity.

It isn't a true grid, so might be hard for people to follow and manipulate.
Users may be be orphaned until someone new connects, which will eventually
give a path form them back to the main group.

The code is now a lot simpler, with a reasonably well factored backend
and no free floating variables.